+++ src/main/java/kappzzang/jeongsan/global/common/security/JwtAuthenticationToken.java
@@ -1,4 +1,4 @@
-package kappzzang.jeongsan.global.security;
+package kappzzang.jeongsan.global.common.security;
 
 import java.util.Collection;
 import org.springframework.security.authentication.AbstractAuthenticationToken;
