+++ src/main/java/kappzzang/jeongsan/global/common/util/JwtProperties.java
@@ -1,4 +1,4 @@
-package kappzzang.jeongsan.global.util;
+package kappzzang.jeongsan.global.common.util;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
