+++ src/main/java/kappzzang/jeongsan/global/exception/GlobalExceptionHandler.java
@@ -1,6 +1,6 @@
 package kappzzang.jeongsan.global.exception;
 
-import kappzzang.jeongsan.global.common.JeongsanApiResponse;
+import kappzzang.jeongsan.global.common.dto.JeongsanApiResponse;
 import kappzzang.jeongsan.global.common.enumeration.ErrorType;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
